+++ BelissimoCloneWPF.Service/Interfaces/Users/IUserService.cs
@@ -1,0 +1,22 @@
+﻿using BelissimoCloneWPF.Domain.Configurations;
+using BelissimoCloneWPF.Domain.Entities.Users;
+using BelissimoCloneWPF.Service.DTOs.Users;
+using System.Linq.Expressions;
+
+namespace BelissimoCloneWPF.Service.Interfaces.Users
+{
+    public interface IUserService
+    {
+        ValueTask<UserForViewDTO> CreateAsync(UserForCreationDTO userForCreationDTO);
+
+        ValueTask<UserForViewDTO> UpdateAsync(int id, UserForUpdateDTO userForUpdateDTO);
+
+        ValueTask<bool> DeleteAsync(int id);
+
+        ValueTask<IEnumerable<UserForViewDTO>> GetAllAsync(
+            PaginationParams @params,
+            Expression<Func<User, bool>> expression = null);
+
+        ValueTask<UserForViewDTO> GetAsync(Expression<Func<User, bool>> expression);
+    }
+}
